+++ src/main/java/org/betonquest/betonquest/config/patcher/migration/migrator/from1to2/FabledRename.java
@@ -1,42 +1,24 @@
 package org.betonquest.betonquest.config.patcher.migration.migrator.from1to2;
 
-import org.betonquest.betonquest.config.patcher.migration.FileConfigurationProvider;
-import org.betonquest.betonquest.config.patcher.migration.Migration;
-import org.bukkit.configuration.file.YamlConfiguration;
-
-import java.io.File;
-import java.io.IOException;
-import java.util.Map;
+import org.betonquest.betonquest.api.bukkit.config.custom.multi.MultiConfiguration;
+import org.betonquest.betonquest.config.patcher.migration.QuestMigration;
+import org.betonquest.betonquest.config.quest.Quest;
 
 /**
  * Handles the fabled rename migration.
  */
-public class FabledRename implements Migration {
-    /**
-     * The config producer.
-     */
-    private final FileConfigurationProvider producer;
+public class FabledRename implements QuestMigration {
 
     /**
      * Creates a new fabled migrator.
-     *
-     * @param provider The config provider
      */
-    public FabledRename(final FileConfigurationProvider provider) {
-        this.producer = provider;
+    public FabledRename() {
     }
 
     @Override
-    public void migrate() throws IOException {
-        final Map<File, YamlConfiguration> configs = producer.getAllConfigs();
-        for (final Map.Entry<File, YamlConfiguration> entry : configs.entrySet()) {
-            final File file = entry.getKey();
-            final YamlConfiguration config = entry.getValue();
-            final boolean cond1Replaced = replaceStartValueInSection(config, "conditions", "skillapiclass", "fabledclass");
-            final boolean cond2Replaced = replaceStartValueInSection(config, "conditions", "skillapilevel", "fabledlevel");
-            if (cond1Replaced || cond2Replaced) {
-                config.save(file);
-            }
-        }
+    public void migrate(final Quest quest) {
+        final MultiConfiguration config = quest.getQuestConfig();
+        replaceStartValueInSection(config, "conditions", "skillapiclass", "fabledclass");
+        replaceStartValueInSection(config, "conditions", "skillapilevel", "fabledlevel");
     }
 }